requirements.txt
reduct-py
influxdb-client[ciso]
aiohttp
pandas
matplotlib
seaborn
numpy
ipykernel